Use double in some places.

+++ src/main/java/proteomics/Index/BuildIndex.java
@@ -126,7 +126,7 @@ public class BuildIndex {
             Set<Short> linkSiteSet = getLinkSiteSet(seq, proteinNTerm);
             if (!linkSiteSet.isEmpty()) {
                 AA[] aaArray = MassTool.seqToAAList(seq);
-                float totalMass = mass_tool_obj.calResidueMass(aaArray) + MassTool.H2O;
+                float totalMass = (float) (mass_tool_obj.calResidueMass(aaArray) + MassTool.H2O);
                 if (totalMass < max_precursor_mass - linker_mass) {
                     int bin = massToBin(totalMass);
                     if (bin_seq_map.containsKey(bin)) {
@@ -147,7 +147,7 @@ public class BuildIndex {
                 linkSiteSet = getLinkSiteSet(varSeq, proteinNTerm);
                 if (!linkSiteSet.isEmpty()) {
                     AA[] aaArray = MassTool.seqToAAList(varSeq);
-                    float totalMass = mass_tool_obj.calResidueMass(aaArray) + MassTool.H2O;
+                    float totalMass = (float) (mass_tool_obj.calResidueMass(aaArray) + MassTool.H2O);
                     if (totalMass < max_precursor_mass - linker_mass) {
                         int bin = massToBin(totalMass);
                         if (bin_seq_map.containsKey(bin)) {
@@ -188,7 +188,7 @@ public class BuildIndex {
                                 int temp_int = mod_free_seq.indexOf('K'); // only use the first K
                                 if ((temp_int != -1) && (temp_int != mod_free_seq.length() - 2)) { // only consider chains whose link site is in the middle.
                                     // consider PTM free
-                                    float mass = mass_tool_obj.calResidueMass(MassTool.seqToAAList(mod_free_seq)) + MassTool.H2O;
+                                    float mass = (float) (mass_tool_obj.calResidueMass(MassTool.seqToAAList(mod_free_seq)) + MassTool.H2O);
                                     if (mass < max_precursor_mass - linker_mass) {
                                         if (uniprot_decoy_mass_seq_map.containsKey(mass)) {
                                             if (uniprot_decoy_mass_seq_map.get(mass).size() < ECL2.random_score_point_t) { // limit the size of set for the sake of memory.
